edms-deployment-20260106-091146/backend/requirements/minimal.txt
# Minimal requirements for basic Django functionality
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
python-decouple==3.8
django-debug-toolbar==4.2.0